Switch to softfp and use correct FreeRTOS port
The Fedora-supplied arm-none-eabi toolchain uses software floating
point. This makes it incompatible with hardfp code. The original
toolchain file I wrote here used float=soft which is entirely
software-based, which sucks on a platform with hardware floating point
support, and also meant I had to use a non-ideal FreeRTOS port. This
commit switches to softfp - where the compiler uses hardware floating
point, but in a way compatible with software floating point libraries.
It's still not absolutely ideal, but will do for now.